answer

+++ 2-selection_sort.c
@@ -19,7 +19,7 @@ void selection_sort(int *array, size_t size)
 			if (array[tmp] > array[j + 1])
 				tmp = j + 1;
 		}
-		if (i != tmp)
+		if (i != (size_t)tmp)
 		{
 			tmp2 = array[i];
 			array[i] = array[tmp];
